+++ build.gradle
@@ -34,6 +34,9 @@ dependencies {
     implementation 'org.slf4j:slf4j-api:1.7.32'
     implementation 'ch.qos.logback:logback-classic:1.4.11'
 
+    // Common libraries
+    implementation 'org.apache.commons:commons-text:1.10.0'
+
     // Testing
     testImplementation 'org.springframework:spring-test:6.0.12'
     testImplementation 'org.junit.jupiter:junit-jupiter:5.10.0'
