+++ src/main/java/com/crm_module/models/training/Training.java
@@ -1,0 +1,24 @@
+package com.crm_module.models.training;
+
+import com.crm_module.models.training.types.TrainingType;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.Duration;
+import java.time.LocalDate;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class Training {
+    private long id;
+    private long traineeId;
+    private long trainerId;
+    private String trainingName;
+    private TrainingType trainingType;
+    private LocalDate trainingDate;
+    private Duration trainingDuration;
+}
