+++ src/main/java/com/crm_module/models/users/User.java
@@ -1,0 +1,18 @@
+package com.crm_module.models.users;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@SuperBuilder
+public abstract class User {
+    private String firstName;
+    private String lastName;
+    private String username;
+    private String password;
+    private boolean isActive;
+}
