+++ src/main/java/com/crm_module/models/users/impl/Trainee.java
@@ -1,0 +1,21 @@
+package com.crm_module.models.users.impl;
+
+import com.crm_module.models.users.User;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+@AllArgsConstructor
+@NoArgsConstructor
+@SuperBuilder
+public class Trainee extends User {
+    private long userId;
+    private String dateOfBirth;
+    private String address;
+}
+
+
