+++ src/main/java/com/crm_module/models/users/impl/Trainer.java
@@ -1,0 +1,20 @@
+package com.crm_module.models.users.impl;
+
+import com.crm_module.models.training.types.TrainingType;
+import com.crm_module.models.users.User;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+@AllArgsConstructor
+@NoArgsConstructor
+@SuperBuilder
+public class Trainer extends User {
+    private long userId;
+    private String specialization;
+    private TrainingType trainingType;
+}
