+++ src/main/java/com/crm_module/services/trainer_service/TrainerService.java
@@ -1,0 +1,13 @@
+package com.crm_module.services.trainer_service;
+
+import com.crm_module.models.users.impl.Trainer;
+import org.springframework.stereotype.Service;
+
+@Service
+public interface TrainerService {
+    Trainer findById(long id);
+
+    Trainer save(Trainer trainer);
+
+    boolean isExistsById(long id);
+}
