+++ src/main/java/com/crm_module/services/training_service/TrainingService.java
@@ -1,0 +1,13 @@
+package com.crm_module.services.training_service;
+
+import com.crm_module.models.training.Training;
+import org.springframework.stereotype.Service;
+
+@Service
+public interface TrainingService {
+    Training findById(long id);
+
+    Training save(Training training);
+
+    boolean isExistsById(long id);
+}
