--- src/main/java/com/crm_module/training/Training.java
@@ -1,23 +1,0 @@
-package com.crm_module.training;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.Duration;
-import java.time.LocalDate;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class Training {
-    private long id;
-    private long traineeId;
-    private long trainerId;
-    private String trainingName;
-    private TrainingType trainingType;
-    private LocalDate trainingDate;
-    private Duration trainingDuration;
-}
