--- src/main/java/com/crm_module/training/TrainingService.java
@@ -1,12 +1,0 @@
-package com.crm_module.training;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface TrainingService {
-    Training findById(long id);
-
-    Training save(Training training);
-
-    boolean isExistsById(long id);
-}
