--- src/main/java/com/crm_module/users/trainee/TraineeService.java
@@ -1,16 +1,0 @@
-package com.crm_module.users.trainee;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface TraineeService {
-    Trainee findById(long id);
-
-    Trainee save(Trainee trainee);
-
-    Trainee update(Trainee trainee);
-
-    boolean deleteById(long id);
-
-    boolean isExistsById(long id);
-}
