--- src/main/java/com/crm_module/users/trainer/TrainerService.java
@@ -1,12 +1,0 @@
-package com.crm_module.users.trainer;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface TrainerService {
-    Trainer findById(long id);
-
-    Trainer save(Trainer trainer);
-
-    boolean isExistsById(long id);
-}
